pyproject.toml
[tool.poetry]
name = "marissa"
version = "0.1.0"
description = "MessAge foRmat Inference with Similarity digeSt Algorithms"
authors = ["Pablo Ruiz"]
readme = "README.md"
packages = [{include = "marissa"}]

[tool.poetry.scripts]
marissa = "marissa.main:app"

[tool.poetry.dependencies]
python = "^3.12"
scapy = "^2.5.0"
typer = {extras = ["all"], version = "^0.9.0"}
py-tlsh = "4.7.2"
pandas = "^2.2.1"
matplotlib = "3.8.2"
kneed = "^0.8.5"
ssdeep = "3.4"
scikit-learn = "^1.4.1.post1"

[tool.poetry.group.dev.dependencies]
black = "^24.4.2"
flake8 = "^7.1.0"
isort = "^5.13.2"

[tool.isort]
profile = "black"
known_first_party = "marissa"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
